+++ packages/browser/test/mocks/simpletransport.ts
@@ -3,7 +3,7 @@ import { BaseTransport } from '../../src/transports';
 
 export class SimpleTransport extends BaseTransport {
   public async sendEvent(_: Event): Promise<Response> {
-    return this.buffer.add(
+    return this._buffer.add(
       Promise.resolve({
         status: Status.fromHttpCode(200),
       }),