+++ packages/hub/src/scope.ts
@@ -10,45 +10,45 @@ import { SyncPromise } from '@sentry/utils/syncpromise';
  */
 export class Scope implements ScopeInterface {
   /** Flag if notifiying is happening. */
-  protected notifyingListeners: boolean = false;
+  protected _notifyingListeners: boolean = false;
 
   /** Callback for client to receive scope changes. */
-  protected scopeListeners: Array<(scope: Scope) => void> = [];
+  protected _scopeListeners: Array<(scope: Scope) => void> = [];
 
   /** Callback list that will be called after {@link applyToEvent}. */
-  protected eventProcessors: EventProcessor[] = [];
+  protected _eventProcessors: EventProcessor[] = [];
 
   /** Array of breadcrumbs. */
-  protected breadcrumbs: Breadcrumb[] = [];
+  protected _breadcrumbs: Breadcrumb[] = [];
 
   /** User */
-  protected user: User = {};
+  protected _user: User = {};
 
   /** Tags */
-  protected tags: { [key: string]: string } = {};
+  protected _tags: { [key: string]: string } = {};
 
   /** Extra */
-  protected extra: { [key: string]: any } = {};
+  protected _extra: { [key: string]: any } = {};
 
   /** Fingerprint */
-  protected fingerprint?: string[];
+  protected _fingerprint?: string[];
 
   /** Severity */
-  protected level?: Severity;
+  protected _level?: Severity;
 
   /**
    * Add internal on change listener. Used for sub SDKs that need to store the scope.
    * @hidden
    */
   public addScopeListener(callback: (scope: Scope) => void): void {
-    this.scopeListeners.push(callback);
+    this._scopeListeners.push(callback);
   }
 
   /**
    * @inheritdoc
    */
   public addEventProcessor(callback: EventProcessor): this {
-    this.eventProcessors.push(callback);
+    this._eventProcessors.push(callback);
     return this;
   }
 
@@ -56,13 +56,13 @@ export class Scope implements ScopeInterface {
    * This will be called on every set call.
    */
   protected notifyScopeListeners(): void {
-    if (!this.notifyingListeners) {
-      this.notifyingListeners = true;
+    if (!this._notifyingListeners) {
+      this._notifyingListeners = true;
       setTimeout(() => {
-        this.scopeListeners.forEach(callback => {
+        this._scopeListeners.forEach(callback => {
           callback(this);
         });
-        this.notifyingListeners = false;
+        this._notifyingListeners = false;
       });
     }
   }
@@ -70,7 +70,7 @@ export class Scope implements ScopeInterface {
   /**
    * This will be called after {@link applyToEvent} is finished.
    */
-  protected notifyEventProcessors(
+  protected _notifyEventProcessors(
     processors: EventProcessor[],
     event: Event | null,
     hint?: EventHint,
@@ -85,10 +85,10 @@ export class Scope implements ScopeInterface {
         const result = processor({ ...event }, hint) as Event | null;
         if (isThenable(result)) {
           (result as Promise<Event | null>)
-            .then(final => this.notifyEventProcessors(processors, final, hint, index + 1).then(resolve))
+            .then(final => this._notifyEventProcessors(processors, final, hint, index + 1).then(resolve))
             .catch(reject);
         } else {
-          this.notifyEventProcessors(processors, result, hint, index + 1)
+          this._notifyEventProcessors(processors, result, hint, index + 1)
             .then(resolve)
             .catch(reject);
         }
@@ -100,7 +100,7 @@ export class Scope implements ScopeInterface {
    * @inheritdoc
    */
   public setUser(user?: User): this {
-    this.user = user ? normalize(user) : {};
+    this._user = user ? normalize(user) : {};
     this.notifyScopeListeners();
     return this;
   }
@@ -109,10 +109,10 @@ export class Scope implements ScopeInterface {
    * @inheritdoc
    */
   public setTags(tags?: { [key: string]: string }): this {
-    this.tags =
+    this._tags =
       tags && isPlainObject(tags)
         ? {
-            ...this.tags,
+            ...this._tags,
             ...normalize(tags),
           }
         : {};
@@ -124,7 +124,7 @@ export class Scope implements ScopeInterface {
    * @inheritdoc
    */
   public setTag(key: string, value: string): this {
-    this.tags = { ...this.tags, [key]: normalize(value) };
+    this._tags = { ...this._tags, [key]: normalize(value) };
     this.notifyScopeListeners();
     return this;
   }
@@ -133,10 +133,10 @@ export class Scope implements ScopeInterface {
    * @inheritdoc
    */
   public setExtras(extra?: { [key: string]: any }): this {
-    this.extra =
+    this._extra =
       extra && isPlainObject(extra)
         ? {
-            ...this.extra,
+            ...this._extra,
             ...normalize(extra),
           }
         : {};
@@ -148,7 +148,7 @@ export class Scope implements ScopeInterface {
    * @inheritdoc
    */
   public setExtra(key: string, extra: any): this {
-    this.extra = { ...this.extra, [key]: normalize(extra) };
+    this._extra = { ...this._extra, [key]: normalize(extra) };
     this.notifyScopeListeners();
     return this;
   }
@@ -157,7 +157,7 @@ export class Scope implements ScopeInterface {
    * @inheritdoc
    */
   public setFingerprint(fingerprint?: string[]): this {
-    this.fingerprint = fingerprint ? normalize(fingerprint) : undefined;
+    this._fingerprint = fingerprint ? normalize(fingerprint) : undefined;
     this.notifyScopeListeners();
     return this;
   }
@@ -166,7 +166,7 @@ export class Scope implements ScopeInterface {
    * @inheritdoc
    */
   public setLevel(level?: Severity): this {
-    this.level = level ? normalize(level) : undefined;
+    this._level = level ? normalize(level) : undefined;
     this.notifyScopeListeners();
     return this;
   }
@@ -178,13 +178,13 @@ export class Scope implements ScopeInterface {
   public static clone(scope?: Scope): Scope {
     const newScope = new Scope();
     Object.assign(newScope, scope, {
-      scopeListeners: [],
+      _scopeListeners: [],
     });
     if (scope) {
-      newScope.extra = { ...scope.extra };
-      newScope.tags = { ...scope.tags };
-      newScope.breadcrumbs = [...scope.breadcrumbs];
-      newScope.eventProcessors = [...scope.eventProcessors];
+      newScope._extra = { ...scope._extra };
+      newScope._tags = { ...scope._tags };
+      newScope._breadcrumbs = [...scope._breadcrumbs];
+      newScope._eventProcessors = [...scope._eventProcessors];
     }
     return newScope;
   }
@@ -193,12 +193,12 @@ export class Scope implements ScopeInterface {
    * @inheritdoc
    */
   public clear(): this {
-    this.breadcrumbs = [];
-    this.tags = {};
-    this.extra = {};
-    this.user = {};
-    this.level = undefined;
-    this.fingerprint = undefined;
+    this._breadcrumbs = [];
+    this._tags = {};
+    this._extra = {};
+    this._user = {};
+    this._level = undefined;
+    this._fingerprint = undefined;
     this.notifyScopeListeners();
     return this;
   }
@@ -207,10 +207,10 @@ export class Scope implements ScopeInterface {
    * @inheritdoc
    */
   public addBreadcrumb(breadcrumb: Breadcrumb, maxBreadcrumbs?: number): this {
-    this.breadcrumbs =
+    this._breadcrumbs =
       maxBreadcrumbs !== undefined && maxBreadcrumbs >= 0
-        ? [...this.breadcrumbs, normalize(breadcrumb)].slice(-maxBreadcrumbs)
-        : [...this.breadcrumbs, normalize(breadcrumb)];
+        ? [...this._breadcrumbs, normalize(breadcrumb)].slice(-maxBreadcrumbs)
+        : [...this._breadcrumbs, normalize(breadcrumb)];
     this.notifyScopeListeners();
     return this;
   }
@@ -219,7 +219,7 @@ export class Scope implements ScopeInterface {
    * @inheritdoc
    */
   public clearBreadcrumbs(): this {
-    this.breadcrumbs = [];
+    this._breadcrumbs = [];
     this.notifyScopeListeners();
     return this;
   }
@@ -228,7 +228,7 @@ export class Scope implements ScopeInterface {
    * Applies fingerprint from the scope to the event if there's one,
    * uses message if there's one instead or get rid of empty fingerprint
    */
-  private applyFingerprint(event: Event): void {
+  private _applyFingerprint(event: Event): void {
     // Make sure it's an array first and we actually have something in place
     event.fingerprint = event.fingerprint
       ? Array.isArray(event.fingerprint)
@@ -237,8 +237,8 @@ export class Scope implements ScopeInterface {
       : [];
 
     // If we have something on the scope, then merge it with event
-    if (this.fingerprint) {
-      event.fingerprint = event.fingerprint.concat(this.fingerprint);
+    if (this._fingerprint) {
+      event.fingerprint = event.fingerprint.concat(this._fingerprint);
     } else if (event.message) {
       // If not, but we have message, use it instead
       event.fingerprint = event.fingerprint.concat(event.message);
@@ -260,27 +260,27 @@ export class Scope implements ScopeInterface {
    * @hidden
    */
   public applyToEvent(event: Event, hint?: EventHint): SyncPromise<Event | null> {
-    if (this.extra && Object.keys(this.extra).length) {
-      event.extra = { ...this.extra, ...event.extra };
+    if (this._extra && Object.keys(this._extra).length) {
+      event.extra = { ...this._extra, ...event.extra };
     }
-    if (this.tags && Object.keys(this.tags).length) {
-      event.tags = { ...this.tags, ...event.tags };
+    if (this._tags && Object.keys(this._tags).length) {
+      event.tags = { ...this._tags, ...event.tags };
     }
-    if (this.user && Object.keys(this.user).length) {
-      event.user = { ...this.user, ...event.user };
+    if (this._user && Object.keys(this._user).length) {
+      event.user = { ...this._user, ...event.user };
     }
-    if (this.level) {
-      event.level = this.level;
+    if (this._level) {
+      event.level = this._level;
     }
 
-    this.applyFingerprint(event);
+    this._applyFingerprint(event);
 
     const hasNoBreadcrumbs = !event.breadcrumbs || event.breadcrumbs.length === 0;
-    if (hasNoBreadcrumbs && this.breadcrumbs.length > 0) {
-      event.breadcrumbs = this.breadcrumbs;
+    if (hasNoBreadcrumbs && this._breadcrumbs.length > 0) {
+      event.breadcrumbs = this._breadcrumbs;
     }
 
-    return this.notifyEventProcessors([...getGlobalEventProcessors(), ...this.eventProcessors], event, hint);
+    return this._notifyEventProcessors([...getGlobalEventProcessors(), ...this._eventProcessors], event, hint);
   }
 }